Summary I've successfully extracted the Strategy pattern for log formatting in ActivityLogger. Here's what was created: New Files Created Core Infrastructure: - LogEntryFormatter.java - Interface defining the formatting contract - AbstractLogEntryFormatter.java - Base class with common utility methods - LogEntryFormatterFactory.java - Factory for obtaining formatters by LogFormat Format-Specific Implementations: - ClfFormatter.java - Common Log Format - ElfFormatter.java - Extended Log Format (NCSA Combined) - W3cFormatter.java - W3C Extended Log File Format - JsonFormatter.java - JSON structured format - LtsvFormatter.java - Labeled Tab-Separated Values - CsvFormatter.java - Comma-Separated Values - SquidFormatter.java - Squid Native access log format - HaproxyFormatter.java - HAProxy HTTP log format - KeyValueFormatter.java - Structured key=value pairs Changes to ActivityLogger - Added LogEntryFormatter formatter field - Initialized formatter in constructor using the factory - Simplified formatLogEntry() to delegate to the formatter - Removed ~700 lines of formatting code (9 format*Entry methods + utility methods) Benefits | Aspect | Improvement | |--------|-------------| | Single Responsibility | Each formatter handles one format | | Testability | Formatters can be tested independently | | Extensibility | Add new format = create one class | | Maintainability | No more giant switch statement | | Reusability | Formatters can be used elsewhere |
