+++ src/main/java/org/littleshoot/proxy/extras/logging/ActivityLogger.java
@@ -6,14 +6,14 @@ import io.netty.handler.codec.http.HttpResponse;
 import java.net.InetSocketAddress;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import javax.net.ssl.SSLSession;
 import org.littleshoot.proxy.ActivityTrackerAdapter;
 import org.littleshoot.proxy.FlowContext;
 import org.littleshoot.proxy.FullFlowContext;
+import org.littleshoot.proxy.extras.logging.formatter.LogEntryFormatter;
+import org.littleshoot.proxy.extras.logging.formatter.LogEntryFormatterFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,6 +40,7 @@ public class ActivityLogger extends ActivityTrackerAdapter {
 
   private final LogFormat logFormat;
   private final LogFieldConfiguration fieldConfiguration;
+  private final LogEntryFormatter formatter;
 
   /** Tracks request timing information. */
   private static class TimedRequest {
@@ -94,6 +95,7 @@ public class ActivityLogger extends ActivityTrackerAdapter {
     this.logFormat = logFormat;
     this.fieldConfiguration =
         fieldConfiguration != null ? fieldConfiguration : LogFieldConfiguration.defaultConfig();
+    this.formatter = LogEntryFormatterFactory.getFormatter(logFormat);
     validateStandardsCompliance();
   }
 
@@ -600,663 +602,22 @@ public class ActivityLogger extends ActivityTrackerAdapter {
   }
 
   /**
-   * Formats a log entry using the configured fields. This method dynamically generates log entries
-   * based on the field configuration rather than hardcoded field lists.
+   * Formats a log entry using the configured formatter. This method delegates to the appropriate
+   * LogEntryFormatter implementation based on the configured log format.
    */
   private String formatLogEntry(
       FlowContext flowContext, TimedRequest timedRequest, HttpResponse response) {
     long duration = System.currentTimeMillis() - timedRequest.startTime;
-
-    StringBuilder sb = new StringBuilder();
     ZonedDateTime now = ZonedDateTime.now(ZoneId.of(UTC));
 
-    switch (logFormat) {
-      case CLF:
-        formatClfEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case ELF:
-        formatElfEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case W3C:
-        formatW3cEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case JSON:
-        formatJsonEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case LTSV:
-        formatLtsvEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case CSV:
-        formatCsvEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case SQUID:
-        formatSquidEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case HAPROXY:
-        formatHaproxyEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-
-      case KEYVALUE:
-        formatKeyValueEntry(sb, flowContext, timedRequest, response, duration, now);
-        break;
-    }
-
-    return sb.toString();
-  }
-
-  /** Formats CLF log entry. */
-  private void formatClfEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    InetSocketAddress clientAddress = flowContext.getClientAddress();
-    String clientIp = clientAddress != null ? clientAddress.getAddress().getHostAddress() : "-";
-
-    // CLF: host ident authuser [date] "request" status bytes
-    sb.append(clientIp).append(" ");
-    sb.append("- "); // ident
-    sb.append("- "); // authuser
-    sb.append("[").append(format(now, DATE_FORMAT_CLF)).append("] ");
-    sb.append("\"")
-        .append(request.method())
-        .append(" ")
-        .append(getFullUrl(request))
-        .append(" ")
-        .append(request.protocolVersion())
-        .append("\" ");
-    sb.append(response.status().code()).append(" ");
-    sb.append(getContentLength(response));
-  }
-
-  /** Formats ELF log entry. */
-  private void formatElfEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    InetSocketAddress clientAddress = flowContext.getClientAddress();
-    String clientIp = clientAddress != null ? clientAddress.getAddress().getHostAddress() : "-";
-
-    // ELF: host ident authuser [date] "request" status bytes "referer" "user-agent"
-    sb.append(clientIp).append(" ");
-    sb.append("- "); // ident
-    sb.append("- "); // authuser
-    sb.append("[").append(format(now, DATE_FORMAT_CLF)).append("] ");
-    sb.append("\"")
-        .append(request.method())
-        .append(" ")
-        .append(getFullUrl(request))
-        .append(" ")
-        .append(request.protocolVersion())
-        .append("\" ");
-    sb.append(response.status().code()).append(" ");
-    sb.append(getContentLength(response)).append(" ");
-    sb.append("\"").append(getHeader(request, "Referer")).append("\" ");
-    sb.append("\"").append(getHeader(request, USER_AGENT)).append("\"");
-  }
-
-  /** Formats W3C log entry. */
-  private void formatW3cEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    InetSocketAddress clientAddress = flowContext.getClientAddress();
-    String clientIp = clientAddress != null ? clientAddress.getAddress().getHostAddress() : "-";
-
-    // W3C: date time c-ip cs-method cs-uri-stem sc-status sc-bytes cs(User-Agent)
-    // cs-uri-stem should be just the path, not full URL
-    DateTimeFormatter w3cDateTimeFormatter =
-        DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss", Locale.US);
-    sb.append(now.format(w3cDateTimeFormatter)).append(" ");
-    sb.append(clientIp).append(" ");
-    sb.append(request.method()).append(" ");
-    sb.append(getUriStem(request)).append(" ");
-    sb.append(response.status().code()).append(" ");
-    sb.append(getContentLength(response)).append(" ");
-    sb.append("\"").append(getHeader(request, USER_AGENT)).append("\"");
-  }
-
-  /** Formats JSON log entry. */
-  private void formatJsonEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    String flowId = timedRequest.flowId;
-
-    sb.append("{");
-    sb.append("\"flow_id\":\"").append(escapeJson(flowId)).append("\"");
-
-    // Use configured fields dynamically
-    boolean first = false;
-    for (LogField field : fieldConfiguration.getFields()) {
-      // Handle prefix-based fields that expand to multiple entries
-      if (field instanceof PrefixRequestHeaderField) {
-        PrefixRequestHeaderField prefixField = (PrefixRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            prefixField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"")
-              .append(entry.getKey())
-              .append("\":\"")
-              .append(escapeJson(entry.getValue()))
-              .append("\"");
-        }
-      } else if (field instanceof PrefixResponseHeaderField) {
-        PrefixResponseHeaderField prefixField = (PrefixResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            prefixField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"")
-              .append(entry.getKey())
-              .append("\":\"")
-              .append(escapeJson(entry.getValue()))
-              .append("\"");
-        }
-      } else if (field instanceof RegexRequestHeaderField) {
-        RegexRequestHeaderField regexField = (RegexRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            regexField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"")
-              .append(entry.getKey())
-              .append("\":\"")
-              .append(escapeJson(entry.getValue()))
-              .append("\"");
-        }
-      } else if (field instanceof RegexResponseHeaderField) {
-        RegexResponseHeaderField regexField = (RegexResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            regexField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"")
-              .append(entry.getKey())
-              .append("\":\"")
-              .append(escapeJson(entry.getValue()))
-              .append("\"");
-        }
-      } else if (field instanceof ExcludeRequestHeaderField) {
-        ExcludeRequestHeaderField excludeField = (ExcludeRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            excludeField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"")
-              .append(entry.getKey())
-              .append("\":\"")
-              .append(escapeJson(entry.getValue()))
-              .append("\"");
-        }
-      } else if (field instanceof ExcludeResponseHeaderField) {
-        ExcludeResponseHeaderField excludeField = (ExcludeResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            excludeField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"")
-              .append(entry.getKey())
-              .append("\":\"")
-              .append(escapeJson(entry.getValue()))
-              .append("\"");
-        }
-      } else {
-        if (!first) {
-          sb.append(",");
-        }
-        first = false;
-
-        String value = field.extractValue(flowContext, request, response, duration);
-        sb.append("\"")
-            .append(field.getName())
-            .append("\":\"")
-            .append(escapeJson(value))
-            .append("\"");
-      }
-    }
-
-    sb.append("}");
-  }
-
-  /** Formats LTSV log entry. */
-  private void formatLtsvEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    String flowId = timedRequest.flowId;
-
-    // Labeled Tab-Separated Values
-    sb.append("flow_id:").append(flowId);
-    boolean first = false;
-    for (LogField field : fieldConfiguration.getFields()) {
-      // Handle prefix-based fields that expand to multiple entries
-      if (field instanceof PrefixRequestHeaderField) {
-        PrefixRequestHeaderField prefixField = (PrefixRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            prefixField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append("\t");
-          }
-          first = false;
-          sb.append(entry.getKey()).append(":").append(entry.getValue());
-        }
-      } else if (field instanceof PrefixResponseHeaderField) {
-        PrefixResponseHeaderField prefixField = (PrefixResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            prefixField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append("\t");
-          }
-          first = false;
-          sb.append(entry.getKey()).append(":").append(entry.getValue());
-        }
-      } else if (field instanceof RegexRequestHeaderField) {
-        RegexRequestHeaderField regexField = (RegexRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            regexField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append("\t");
-          }
-          first = false;
-          sb.append(entry.getKey()).append(":").append(entry.getValue());
-        }
-      } else if (field instanceof RegexResponseHeaderField) {
-        RegexResponseHeaderField regexField = (RegexResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            regexField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append("\t");
-          }
-          first = false;
-          sb.append(entry.getKey()).append(":").append(entry.getValue());
-        }
-      } else if (field instanceof ExcludeRequestHeaderField) {
-        ExcludeRequestHeaderField excludeField = (ExcludeRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            excludeField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append("\t");
-          }
-          first = false;
-          sb.append(entry.getKey()).append(":").append(entry.getValue());
-        }
-      } else if (field instanceof ExcludeResponseHeaderField) {
-        ExcludeResponseHeaderField excludeField = (ExcludeResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            excludeField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append("\t");
-          }
-          first = false;
-          sb.append(entry.getKey()).append(":").append(entry.getValue());
-        }
-      } else {
-        if (!first) {
-          sb.append("\t");
-        }
-        first = false;
-
-        String value = field.extractValue(flowContext, request, response, duration);
-        sb.append(field.getName()).append(":").append(value);
-      }
-    }
-  }
-
-  /** Formats CSV log entry. */
-  private void formatCsvEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    String flowId = timedRequest.flowId;
-
-    // Comma-Separated Values
-    sb.append("\"").append(escapeJson(flowId)).append("\"");
-    boolean first = false;
-    for (LogField field : fieldConfiguration.getFields()) {
-      // Handle prefix-based fields that expand to multiple entries
-      if (field instanceof PrefixRequestHeaderField) {
-        PrefixRequestHeaderField prefixField = (PrefixRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            prefixField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"").append(escapeJson(entry.getValue())).append("\"");
-        }
-      } else if (field instanceof PrefixResponseHeaderField) {
-        PrefixResponseHeaderField prefixField = (PrefixResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            prefixField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"").append(escapeJson(entry.getValue())).append("\"");
-        }
-      } else if (field instanceof RegexRequestHeaderField) {
-        RegexRequestHeaderField regexField = (RegexRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            regexField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"").append(escapeJson(entry.getValue())).append("\"");
-        }
-      } else if (field instanceof RegexResponseHeaderField) {
-        RegexResponseHeaderField regexField = (RegexResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            regexField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"").append(escapeJson(entry.getValue())).append("\"");
-        }
-      } else if (field instanceof ExcludeRequestHeaderField) {
-        ExcludeRequestHeaderField excludeField = (ExcludeRequestHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            excludeField.extractMatchingHeaders(request.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"").append(escapeJson(entry.getValue())).append("\"");
-        }
-      } else if (field instanceof ExcludeResponseHeaderField) {
-        ExcludeResponseHeaderField excludeField = (ExcludeResponseHeaderField) field;
-        for (Map.Entry<String, String> entry :
-            excludeField.extractMatchingHeaders(response.headers()).entrySet()) {
-          if (!first) {
-            sb.append(",");
-          }
-          first = false;
-          sb.append("\"").append(escapeJson(entry.getValue())).append("\"");
-        }
-      } else {
-        if (!first) {
-          sb.append(",");
-        }
-        first = false;
-
-        String value = field.extractValue(flowContext, request, response, duration);
-        sb.append("\"").append(escapeJson(value)).append("\"");
-      }
-    }
-  }
-
-  /** Formats SQUID log entry. */
-  private void formatSquidEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    InetSocketAddress clientAddress = flowContext.getClientAddress();
-    String clientIp = clientAddress != null ? clientAddress.getAddress().getHostAddress() : "-";
-
-    // Squid format: time elapsed remotehost code/status bytes method URL rfc931 peerstatus/peerhost
-    // type
-    // Time is seconds since epoch with milliseconds (e.g., 1234567890.123)
-    long epochSeconds = now.toEpochSecond();
-    long millis = now.getNano() / 1_000_000;
-    sb.append(epochSeconds).append(".").append(String.format("%03d", millis)).append(" ");
-    sb.append(duration).append(" "); // elapsed time in ms
-    sb.append(clientIp).append(" ");
-    // Determine cache result code based on response
-    String cacheResult = determineCacheResult(response);
-    sb.append(cacheResult).append("/").append(response.status().code()).append(" ");
-    sb.append(getContentLength(response)).append(" ");
-    sb.append(request.method()).append(" ");
-    sb.append(getFullUrl(request)).append(" ");
-    sb.append("- "); // rfc931 (ident)
-    sb.append("DIRECT/").append(getServerIp(flowContext)).append(" ");
-    sb.append(getContentType(response));
-  }
-
-  /**
-   * Determines the cache result code for Squid format based on response. Returns TCP_HIT for
-   * 2xx/3xx status codes, TCP_MISS for others.
-   *
-   * @param response the HTTP response
-   * @return cache result code (TCP_HIT or TCP_MISS)
-   */
-  private String determineCacheResult(HttpResponse response) {
-    int status = response.status().code();
-    // 2xx and 3xx are generally considered cacheable hits
-    if (status >= 200 && status < 400) {
-      return "TCP_HIT";
-    }
-    return "TCP_MISS";
-  }
-
-  /** Formats HAProxy log entry. */
-  private void formatHaproxyEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    InetSocketAddress clientAddress = flowContext.getClientAddress();
-    String clientIp = clientAddress != null ? clientAddress.getAddress().getHostAddress() : "-";
-    int clientPort = clientAddress != null ? clientAddress.getPort() : 0;
-
-    // HAProxy HTTP log format:
-    // process[pid]: client_ip:port [accept_date] frontend backend/server Tq Tw Tc Tr Ta status
-    // bytes
-    // cc cs sc rc sr st rt request
-    // Since LittleProxy doesn't have frontend/backend/server separation, we use simplified format
-    sb.append("littleproxy[0]: "); // process name and pid
-    sb.append(clientIp).append(":").append(clientPort).append(" ");
-    sb.append("[").append(format(now, "dd/MMM/yyyy:HH:mm:ss.SSS")).append("] ");
-    sb.append("frontend backend/server "); // placeholder for frontend/backend/server
-    sb.append("0/0/0/0/").append(duration).append(" "); // Tq/Tw/Tc/Tr/Ta timing
-    sb.append(response.status().code()).append(" ");
-    sb.append(getContentLength(response)).append(" ");
-    sb.append("- - - - - - - "); // cc cs sc rc sr st rt (counters and termination state)
-    sb.append("\"")
-        .append(request.method())
-        .append(" ")
-        .append(getFullUrl(request))
-        .append(" ")
-        .append(request.protocolVersion())
-        .append("\"");
-  }
-
-  /** Formats KEYVALUE log entry (structured text format). */
-  private void formatKeyValueEntry(
-      StringBuilder sb,
-      FlowContext flowContext,
-      TimedRequest timedRequest,
-      HttpResponse response,
-      long duration,
-      ZonedDateTime now) {
-    HttpRequest request = timedRequest.request;
-    InetSocketAddress clientAddress = flowContext.getClientAddress();
-    String clientIp = clientAddress != null ? clientAddress.getAddress().getHostAddress() : "-";
-    int clientPort = clientAddress != null ? clientAddress.getPort() : 0;
-
-    // Get client state metrics
-    ClientState clientState = clientStates.get(clientAddress);
-    long clientConnectTime = clientState != null ? clientState.connectTime : 0;
-    long clientConnectionDuration =
-        clientConnectTime > 0 ? System.currentTimeMillis() - clientConnectTime : 0;
-    long sslHandshakeTime =
-        clientState != null && clientState.sslHandshakeEndTime > 0
-            ? clientState.sslHandshakeEndTime - clientState.sslHandshakeStartTime
-            : 0;
-    int clientSaturations = clientState != null ? clientState.saturationCount : 0;
-    String exceptionType = clientState != null ? clientState.lastExceptionType : null;
-
-    // Get server state metrics
-    ServerState serverState =
-        flowContext instanceof FullFlowContext
-            ? serverStates.get((FullFlowContext) flowContext)
-            : null;
-    long serverConnectTime =
-        serverState != null && serverState.connectEndTime > 0
-            ? serverState.connectEndTime - serverState.connectStartTime
-            : 0;
-    int serverSaturations = serverState != null ? serverState.saturationCount : 0;
-    String serverAddress =
-        serverState != null && serverState.remoteAddress != null
-            ? serverState.remoteAddress.getAddress().getHostAddress()
-            : "-";
-    int serverPort =
-        serverState != null && serverState.remoteAddress != null
-            ? serverState.remoteAddress.getPort()
-            : 0;
-
-    // Build structured log entry: flow_id=... client_ip=... key=value pairs
-    sb.append("flow_id=").append(timedRequest.flowId);
-    sb.append(" client_ip=").append(clientIp);
-    sb.append(" client_port=").append(clientPort);
-    sb.append(" server_ip=").append(serverAddress);
-    sb.append(" server_port=").append(serverPort);
-    sb.append(" method=").append(request.method().name());
-    sb.append(" uri=\"").append(request.uri()).append("\"");
-    sb.append(" protocol=").append(request.protocolVersion());
-    sb.append(" status=").append(response.status().code());
-    sb.append(" bytes=").append(getContentLength(response));
-    sb.append(" http_request_ms=").append(duration);
-    sb.append(" tcp_connection_ms=").append(clientConnectionDuration);
-    sb.append(" server_connect_ms=").append(serverConnectTime);
-    sb.append(" ssl_handshake_ms=").append(sslHandshakeTime);
-    sb.append(" client_saturations=").append(clientSaturations);
-    sb.append(" server_saturations=").append(serverSaturations);
-    sb.append(" exception=").append(exceptionType != null ? exceptionType : "none");
-  }
-
-  /**
-   * Reconstructs the full URL from the request. If the URI is already absolute (starts with http://
-   * or https://), returns it as-is. Otherwise, prepends the Host header to create a complete URL.
-   *
-   * @param request the HTTP request
-   * @return the full URL
-   */
-  protected String getFullUrl(HttpRequest request) {
-    String uri = request.uri();
-
-    // Check if URI is already absolute (contains scheme)
-    if (uri.startsWith("http://") || uri.startsWith("https://")) {
-      return uri;
-    }
-
-    // For CONNECT requests, the URI is just host:port
-    if (request.method().name().equals("CONNECT")) {
-      return uri;
-    }
-
-    // Get host from Host header
-    String host = request.headers().get("Host");
-    if (host == null || host.isEmpty()) {
-      // Fallback: return URI as-is if no Host header
-      return uri;
-    }
-
-    // Determine scheme (default to http)
-    String scheme = "http";
-
-    // Reconstruct full URL
-    if (uri.startsWith("/")) {
-      return scheme + "://" + host + uri;
-    } else {
-      return scheme + "://" + host + "/" + uri;
-    }
-  }
-
-  /**
-   * Extracts the URI stem (path) from the request for W3C format. Returns just the path component
-   * without scheme and host.
-   *
-   * @param request the HTTP request
-   * @return the URI stem (path)
-   */
-  protected String getUriStem(HttpRequest request) {
-    String uri = request.uri();
-
-    // If it's already a path (starts with /), return as-is
-    if (uri.startsWith("/")) {
-      return uri;
-    }
-
-    // If it's a full URL, extract just the path
-    if (uri.startsWith("http://") || uri.startsWith("https://")) {
-      int pathStart = uri.indexOf("/", uri.indexOf("://") + 3);
-      if (pathStart == -1) {
-        return "/";
-      }
-      int queryStart = uri.indexOf("?", pathStart);
-      if (queryStart == -1) {
-        return uri.substring(pathStart);
-      } else {
-        return uri.substring(pathStart, queryStart);
-      }
-    }
-
-    // For CONNECT requests or other cases, return as-is
-    return uri;
-  }
-
-  /**
-   * Gets header value from request.
-   *
-   * @param request the HTTP request
-   * @param headerName the header name
-   * @return header value or "-"
-   */
-  protected String getHeader(HttpRequest request, String headerName) {
-    String val = request.headers().get(headerName);
-    return val != null ? val : "-";
+    return formatter.format(
+        flowContext,
+        timedRequest.request,
+        response,
+        duration,
+        now,
+        timedRequest.flowId,
+        fieldConfiguration);
   }
 
   /**
@@ -1268,57 +629,5 @@ public class ActivityLogger extends ActivityTrackerAdapter {
   protected String getContentLength(HttpResponse response) {
     String len = response.headers().get("Content-Length");
     return len != null ? len : "-";
-  }
-
-  /**
-   * Gets content type from response.
-   *
-   * @param response the HTTP response
-   * @return content type or "-"
-   */
-  protected String getContentType(HttpResponse response) {
-    String val = response.headers().get("Content-Type");
-    return val != null ? val : "-";
-  }
-
-  /**
-   * Gets server IP from flow context.
-   *
-   * @param context the flow context
-   * @return server IP or "-"
-   */
-  protected String getServerIp(FlowContext context) {
-    if (context instanceof FullFlowContext) {
-      String hostAndPort = ((FullFlowContext) context).getServerHostAndPort();
-      if (hostAndPort != null) {
-        // Returns "host:port", we want just the host/ip usually, or stick with
-        // host:port as Squid format usually asks for remotehost or peerhost
-        return hostAndPort.split(":")[0];
-      }
-    }
-    return "-";
-  }
-
-  /**
-   * Escapes JSON strings for safe logging.
-   *
-   * @param s the string to escape
-   * @return escaped string
-   */
-  protected String escapeJson(String s) {
-    if (s == null) return "";
-    return s.replace("\"", "\\\"").replace("\\", "\\\\");
-  }
-
-  /**
-   * Formats timestamp using specified pattern.
-   *
-   * @param zonedDateTime the date/time to format
-   * @param pattern the date format pattern
-   * @return formatted timestamp
-   */
-  private String format(ZonedDateTime zonedDateTime, String pattern) {
-    DateTimeFormatter dtf = DateTimeFormatter.ofPattern(pattern, Locale.US);
-    return zonedDateTime.format(dtf);
   }
 }
